Changes to the Client Library
To Allow for Setting of the page size when querying. You can now
set this through the sue of a System property, "recurly.page.size".
Pagiantion is not supported but this is something we shold look into.

Subscriptions can now also be queried for a given user.

You can dump the XML messgaes passing between Recurly and the client
through the use of a System Property "recurly.debug".