Refactor and test code around equals and hashCode
This code ensures that each model overrides `hashCode()`. It also makes
sure not to use super's hashCode implementation because that was
including references. We want to do a logical compare and not a
reference compare. This code also removes the individual overrides of
equals in favor of sharing one equals method on the `RecurlyObject` class.
It also adds more test cases for these methods and a way to
deterministically build random model data.